notification_microservice/microservice/requirements.txt
fastapi==0.91.0
pydantic==1.10.4
requests==2.28.2
uvicorn==0.20.0
pandas==2.0.1
kafka-python==2.0.2
psycopg2-binary==2.9.6